Refactored componets directory structure and added a BookInfo component that will display a specific books information
